fix(recording): keep the seeded obj_nearby counter unsigned (#232)
`NextNameIndex` clamps at 0xFFFFFFFE precisely so that "highest + 1" cannot
wrap and hand back a name already defined. Narrowing the result to `int` at
the call site threw that guarantee away: 0xFFFFFFFF becomes -1, the first
name becomes `obj_nearby-1` and the second `obj_nearby0`, which may well be
in use — and incrementing past INT_MAX is undefined besides.

`int` was right when the counter started at 0 and only counted this
session's edges. It is not right now that it starts wherever the loaded pack
left off, which is a value the pack's author controls.

Only reachable by hand-editing a name into the manifest, so no case changes;
776/776 still pass.

+++ Core/NES/HdPacks/HdPackBuilder.cpp
@@ -309,7 +309,11 @@ void HdPackBuilder::BuildObjectSheets(stringstream& tileRows)
 		candidates.push_back(adjacency);
 	}
 
-	int edgeIndex = (int)MesenSheets::NextNameIndex(loadedConditionNames, "obj_nearby");
+	//uint32_t, not int: the counter no longer starts at 0, it starts wherever the
+	//loaded pack left off. Narrowing that to int would turn NextNameIndex's clamp
+	//into a negative index - and the name it guarantees is free into one that may
+	//not be - and would make incrementing past INT_MAX undefined.
+	uint32_t edgeIndex = MesenSheets::NextNameIndex(loadedConditionNames, "obj_nearby");
 	for(size_t index : MesenSheets::SelectTileNearby(candidates, kTileNearbyMinFrames, kTileNearbyMinProbability)) {
 		const MesenSheets::TileAdjacency& edge = candidates[index];
 		uint32_t a = edge.A, b = edge.B;
